+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Removed
+- **Breaking change for Consumer API**: Removed `processed_at` timestamp for material requirements
 - **Breaking change for Supplier API**: Removed MaterialAvailabilityCheck for suppliers
 
 ## [0.4.0]