+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.1] - 2022-02-10
 ### Added
  - Added distributors to suppliers
 
@@ -53,7 +54,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Added support for material requirements
 
-[Unreleased]: https://github.com/zaikio/zaikio-procurement-ruby/compare/v0.4.0..HEAD
+[Unreleased]: https://github.com/zaikio/zaikio-procurement-ruby/compare/v0.4.1..HEAD
+[0.4.1]: https://github.com/zaikio/zaikio-procurement-ruby/compare/v0.4.0..v0.4.1
 [0.4.0]: https://github.com/zaikio/zaikio-procurement-ruby/compare/v0.3.2..v0.4.0
 [0.3.2]: https://github.com/zaikio/zaikio-procurement-ruby/compare/v0.3.1..v0.3.2
 [0.3.1]: https://github.com/zaikio/zaikio-procurement-ruby/compare/v0.2.2..v0.3.1